requirements.txt
flask
PyMuPDF 
openai
nltk 
Markup
python-dotenv
azure.cognitiveservices.vision.computervision